+++ R/utils.R
@@ -146,3 +146,20 @@ cov_type <- c("us", "toep", "toeph", "ar1", "ar1h", "ad", "adh", "cs", "csh")
 #' @describeIn covariance_types spatial covariance structure
 #' @format NULL
 cov_type_spatial <- c("sp_exp")
+
+
+#' Trace of a Matrix
+#'
+#' @description Obtain the trace of a matrix if the matrix is diagonal, otherwise raise an error.
+#'
+#' @param x (`matrix`)\cr square matrix input.
+#'
+#' @return The trace of the square matrix.
+#'
+#' @keywords internal
+h_tr <- function(x) {
+  if (nrow(x) != ncol(x)) {
+    stop("x must be square matrix")
+  }
+  sum(diag(x))
+}